% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleEngine.R
\name{RuleEngine}
\alias{RuleEngine}
\title{RuleEngine}
\usage{
RuleEngine(df, unique_identifier_column, database_path = ":memory:")
}
\arguments{
\item{df}{Dataframe object in which you want to process the indicators.}

\item{unique_identifier_column}{Name of the column containing the unique identifiers for the provided dataframe (there can be no repeated values in this column).}

\item{database_path}{Path where you want to save the database needed to calculate the indicators (by default in memory).}
}
\value{
A Rule Engine Object.
}
\description{
RuleEngine
}
\examples{
\dontrun{
df <- read.csv("dataset.csv", sep = "|", header = TRUE)
reng <- rule_engine(df,"hospitalization_id")

df2 <- read.csv("dataset2.csv", sep = "|", header = TRUE)
reng2 <- rule_engine(df2,"episode_id","./indicators.duckdb")
}
}
