% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruleEngine.R
\docType{class}
\name{SqlRuleIndicator-class}
\alias{SqlRuleIndicator-class}
\alias{SqlRuleIndicator}
\title{Create Parent Class SqlRuleIndicator}
\arguments{
\item{name}{Name of the indicator.}

\item{sql_rule}{Logic of the indicator.}
}
\value{
An SqlRuleIndicator Object.
}
\description{
Create Parent Class SqlRuleIndicator
}

